Simplify type conversions

+++ src/gui/SampleThumbnail.cpp
@@ -115,7 +115,7 @@ void SampleThumbnail::visualize(VisualizeParameters parameters, QPainter& painte
 	const auto sampleRange = parameters.sampleEnd - parameters.sampleStart;
 	if (sampleRange <= 0.0f || sampleRange > 1.0f) { return; }
 
-	const auto targetThumbnailWidth = static_cast<int>(static_cast<double>(sampleRect.width()) / sampleRange);
+	const auto targetThumbnailWidth = static_cast<int>(sampleRect.width() / sampleRange);
 	const auto finerThumbnail = std::find_if(m_thumbnailCache->rbegin(), m_thumbnailCache->rend(),
 		[&](const auto& thumbnail) { return thumbnail.width() >= targetThumbnailWidth; });
 
@@ -128,8 +128,8 @@ void SampleThumbnail::visualize(VisualizeParameters parameters, QPainter& painte
 	painter.save();
 	painter.setRenderHint(QPainter::Antialiasing, true);
 
-	const auto thumbnailBeginForward = std::max<int>(renderRect.x() - sampleRect.x(), static_cast<int>(parameters.sampleStart * targetThumbnailWidth));
-	const auto thumbnailEndForward = std::max<int>(renderRect.x() + renderRect.width() - sampleRect.x(), static_cast<int>(parameters.sampleEnd * targetThumbnailWidth));
+	const auto thumbnailBeginForward = std::max<int>(renderRect.x() - sampleRect.x(), parameters.sampleStart * targetThumbnailWidth);
+	const auto thumbnailEndForward = std::max<int>(renderRect.x() + renderRect.width() - sampleRect.x(), parameters.sampleEnd * targetThumbnailWidth);
 	const auto thumbnailBegin = parameters.reversed ? targetThumbnailWidth - thumbnailBeginForward - 1 : thumbnailBeginForward;
 	const auto thumbnailEnd = parameters.reversed ? targetThumbnailWidth - thumbnailEndForward : thumbnailEndForward;
 	const auto advanceThumbnailBy = parameters.reversed ? -1 : 1;
@@ -140,8 +140,8 @@ void SampleThumbnail::visualize(VisualizeParameters parameters, QPainter& painte
 	for (auto x = renderRect.x(), i = thumbnailBegin; x < renderRect.x() + renderRect.width() && i != thumbnailEnd;
 		 ++x, i += advanceThumbnailBy)
 	{
-		const auto beginAggregationAt = &(*finerThumbnail)[static_cast<int>(std::floor(i * finerThumbnailScaleFactor))];
-		const auto endAggregationAt = &(*finerThumbnail)[static_cast<int>(std::ceil((i + 1) * finerThumbnailScaleFactor))];
+		const auto beginAggregationAt = &(*finerThumbnail)[static_cast<size_t>(std::floor(i * finerThumbnailScaleFactor))];
+		const auto endAggregationAt = &(*finerThumbnail)[static_cast<size_t>(std::ceil((i + 1) * finerThumbnailScaleFactor))];
 		const auto peak = std::accumulate(beginAggregationAt, endAggregationAt, Thumbnail::Peak{});
 
 		const auto yMin = renderRect.center().y() - peak.min * yScale;
